[Nexus] Corrected form of #5624
This fixes the problems introduced in #5624 by reverting the modifications to `testing.py` which previously caused all Nexus tests to simply not be run. In addition, the changes mentioned in #5624 were implemented, however the change for `pseudopotential.py` was modified to use `np.minimum` and `np.maximum` which are, shockingly, different functions than `np.min` and `np.max`.