[Nexus] Potential replacement for `plotting.py`
This is a potential way that I might go about removing `plotting.py` from Nexus, pending approval.

The changes in `gaussian_process.py` are small and simple, shouldn't be much contention over that.

The changes to `labs/lab2_qmc_basics/oxygen_atom/ip_conv.py` on the other hand are larger and change some small bits of how the code works. In general everything remaings the same, except for how the figures are made and how things are plotted on the figures.