[Nexus] Protect against failed Matplotlib import in lab
This brings back the protection that `plotting.py` had for a failed import of Matplotlib by just nesting the whole piece of plotting in a `try... except` block. It specifically is designed to only catch `ImportError` since any other error that may arise would be unrelated to the Matplotlib import, and so should get reported.